requirements.txt
aiohttp==2.3.2
aiomysql==0.0.11
appnope==0.1.0
async-timeout==2.0.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.3
blessings==1.6.1
bpython==0.17.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
curtsies==0.3.0
cycler==0.10.0
decorator==4.3.0
Django==2.0
entrypoints==0.2.3
Flask==0.12.2
greenlet==0.4.13
html5lib==0.999999999
idna==2.6
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
itsdangerous==0.24
jedi==0.12.0
jieba==0.39
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
multidict==3.3.2
mysql-connector==2.1.4
mysql-connector-python==8.0.6
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.5.0
numpy==1.13.3
olefile==0.44
pandas==0.21.0
pandocfilters==1.4.2
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
PyMySQL==0.8.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.1.14
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.2.1
yarl==0.13.0
